refact : <최기성 #12> Test 코드 리팩토링
1. Test 코드 리팩토링
