+++ src/main/java/com/example/jariBean/dto/reserved/ReservedReqDto.java
@@ -38,7 +38,7 @@ public class ReservedReqDto {
         private String userId;
 
         @NotEmpty(message = "유저의 시간은 필수입니다.")
-        private String userNow;
+        private LocalDateTime userNow;
 
     }
